pyproject.toml
[project]
name = "observability-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "sqlalchemy>=2.0.29",
    "clickhouse-sqlalchemy>=0.3.0",
    "clickhouse-driver>=0.2.7",
    "python-dotenv>=1.0.1",
    "langgraph>=0.0.39",
    "langchain-groq>=0.1.3",
    "langchain-openai>=0.1.1",
    "langchain-core>=0.1.52",
]

[project.optional-dependencies]
eval = [
    "ragas>=0.1.0",
    "deepeval>=0.21.0",
    "arize-phoenix-evals",
]

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true

